Extend block device data types
The size of large block devices overflows when these variables are 32 bits long. For example, a 32 GB SD Card is reported as smaller than it really is.

+++ libraries/BlockDevices/BlockDevice.h
@@ -30,8 +30,8 @@
 #define  BD_ERROR_DEVICE_ERROR  (-4001)
 #define  BD_ERROR_WRITE_PROTECTED (4002)
 
-typedef uint32_t bd_addr_t;
-typedef uint32_t bd_size_t;
+typedef uint64_t bd_addr_t;
+typedef uint64_t bd_size_t;
 typedef pin_size_t pin_t;
 
 /* -------------------------------------------------------------------------- */
